refactor Character type and update files
