+++ constants/types/Character.ts
@@ -2,4 +2,5 @@ export type Character = {
   name: string;
   color: string;
   icon: string;
+  iconIndex: number;
 };
